feat/699/EndPointPost/PermissionsImp-FixDeErrores
