--- AuthMicroservice Backend/src/main/java/upb/edu/AuthMicroservice/models/Permissions.java
@@ -1,19 +1,0 @@
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/permissions")
-public class PermissionController {
-
-    private final CreatePermissionInteractor interactor;
-
-    public PermissionController(CreatePermissionInteractor interactor) {
-        this.interactor = interactor;
-    }
-
-    @PostMapping
-    public ResponseEntity<BaseResponse> createPermission(@RequestBody CreatePermissionRequest request) {
-        BaseResponse response = interactor.execute(request);
-        return ResponseEntity.status(201).body(response);
-    }
-}
